+++ src/main.rs
@@ -121,10 +121,10 @@ pub fn run_opt(opt: &Opt) -> Result<bool, Error> {
         ret
     } else {
         println!(
-            "Config file '{}' is not found. Use default config",
+            "Config file '{}' is not found. Enable all rules",
             opt.config.to_string_lossy()
         );
-        Config::new()
+        Config::new().enable_all()
     };
 
     run_opt_config(opt, config)